requirements.txt
pyyaml
beautifulsoup4
lxml
